fix errant envelope resetting behavior
calls to update() are a bit of a mess. properly reset an envelope when the patch is updated, but not in the normal case of a key down or key up!

allow setting of two midi control IDs to use as expression control source (default modulation wheel and expression control). Could be set to 2 for breath control, for example.

support aftertouch. We assume no single keyboard will be sending _both_ polyphonic and mono aftertouch, but perhaps we should prepare for this to happen.