fix use after free
it's a stupid one on my part; op_patch needs to be a shared pointer as just returning raw internal memory from the state instance will not work (and in fact _did not_ work). As its owned by the patch which itself goes through the ptr_msg, it will be released when the patch is so it can just be a plain shared_ptr used by everyone.