re-assert some patterns
when accessing shared_ptr objects from the patch state, there's no reason for them not to be const reference. put that back.

although the op_patch is now a shared pointer, its lifetime is at least that of the patch it came from, so put the pattern used for patch components back in the op. We can safely use op_ptr::get() within the context of our patch from the ptr_msg.